nydusd: move shared structures to a lib
Nydusctl will use those structures to deserialize http
responses. So showing nydusd metrics and running status is
more eaiser.
